Moving to Pattern-Power for pattern matching.
The actual pattern matching code has been abstracted into the
Pattern-Power (!!) library, available at colah/Pattern-Power.

Suddenly it became possible to delete a lot of code!

The pattern matchng code is now type checked in its construction, as
apposed to earlier where it was TH constructed and checked when the
quasiquoter was executed at compile time. This makes bugs easier to find
and fix.

A nuber of bugs that actually boiled down to pattern matching fail have
been fixed.